+++ packages/package_info/CHANGELOG.md
@@ -1,3 +1,8 @@
+## 0.2.0
+
+* **Breaking change**. Introduced class `PackageInfo` in place of individual functions.
+* `PackageInfo` provides all package information with a single async call.
+
 ## 0.1.1
 
 * Added package name to available information.